Add files via upload
This file contains the panel dataset for  400 counties, extending from 1990 to 2020, and it uses the economic data provided by the Eurostat database. The economic data I have included are: 
1. gdppc - gdp per capital, at constant prices of 2015
2. total employment
3. real labour productivity per worker
4. consumption of fixed capital

the climate data remains the same, with tas (temperature at surface), prec (precipitation), precpd (precipitation intensity), hotdays (number of extremely hot days in a year) and colddays ( number of extrememly colddays in a year).

Additionally, the climate data has information on hotdays, colddays and precpd at seasonal level. This seasonal definition is indicated by
1.  *** <season>_sn => winter is dec, jan and feb, and so there is year overlap from previous to next *** 
                                                                        and 
2. *** <season>_sq => where seasons are contained within each year***